### 2020-10-06 #2724 @whedon
[PRE REVIEW]: rdataretriever: An R package for downloading, cleaning, and installing publicly available datasets
**Submitting author:** @henrykironde (<a href="http://orcid.org/0000-0001-7105-5808">henry senyondo</a>)
**Repository:** <a href="https://github.com/ropensci/rdataretriever" target ="_blank">https://github.com/ropensci/rdataretriever</a>
**Version:** v3.0.0
