### 2020-10-06 #2724 @kthyng
Hi @henrykironde! Thank you for your submission. I have flagged your submission as possibly out of scope due to its small size. Please see details on "substantial scholarly effort": https://joss.readthedocs.io/en/latest/review_criteria.html#substantial-scholarly-effort. The editorial board will consider your submission to make the decision together in the next week or two. If there is any additional information you'd like to provide, please feel free to do so in this pre-review issue.
