### 2020-10-06 #2724 @ethanwhite
Thanks for the update @kthyng. FYI we did submit a pre-submission inquiry about a month ago:
https://github.com/openjournals/joss/issues/796
