+++ man/install_hdf5.Rd
@@ -1,0 +1,42 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/rdataretriever.R
+\name{install_hdf5}
+\alias{install_hdf5}
+\title{Install datasets via the Data Retriever.}
+\usage{
+install_hdf5(
+  dataset,
+  file = "hdf5.h5",
+  table_name = "{db}_{table}",
+  data_dir = getwd(),
+  debug = FALSE,
+  use_cache = TRUE,
+  force = FALSE,
+  hash_value = NULL
+)
+}
+\arguments{
+\item{dataset}{the name of the dataset that you wish to install or path to a committed dataset zip file}
+
+\item{file}{hdf5 database file name or path}
+
+\item{table_name}{table name for installing of dataset}
+
+\item{data_dir}{the dir path to store the db, defaults to working dir}
+
+\item{debug}{Setting TRUE helps in debugging in case of errors}
+
+\item{use_cache}{Setting FALSE reinstalls scripts even if they are already installed}
+
+\item{force}{Setting TRUE doesn't prompt for confirmation while installing committed datasets when changes are discovered in environment}
+
+\item{hash_value}{the hash value of committed dataset when installing from provenance directory}
+}
+\description{
+Data is stored in hdf5 database
+}
+\examples{
+\donttest{
+rdataretriever::install_hdf5(dataset = "iris", file = "hdf5.h5")
+}
+}
