### 2020-10-07 #2724 @kthyng
@ethanwhite Thank you for including that information! We'll use it as part of our decision-making.
