### 2020-10-14 #2724 @danielskatz
@henrykironde & @ethanwhite - Has this software gone through rOpenSci review?
