### 2020-10-14 #2724 @henrykironde
It is under rOpenSci, so I believe it was reviewed during the early versions before I became the maintainer. `https://github.com/ropensci/rdataretriever`
