### 2020-10-15 #2724 @ethanwhite
@danielskatz - the project's involvement in rOpenSci predates the creation of the rOpenSci review process so it has not formally been through there review process.
