### 2020-10-15 #2724 @fboehm
@henrykironde - would you be willing to write a vignette for the R package? I'm interested in seeing the package used in practice to access an interesting data set. How do I, as a user, know which data sets are available? Maybe a vignette could list the currently available data sets and demonstrate something interesting with one or two data sets. Thank you!
