### 2020-10-19 d7e4438
Add vignette of analysis using BBS data
