### 2020-10-25 27d5687
Remove package installation since this doesn't belong in vignette
