### 2020-10-26 2414108
Add first draft of BBS vignette (#255)
* Add vignette of analysis using BBS data

Move vignette folder to proper location
Remove package installation since this doesn't belong in vignette
