+++ vignettes/breed-bird-survey-analysis.Rmd
@@ -1,0 +1,149 @@
+---
+title: "Using the rdataretriever to quickly analyze Breeding Bird Survey data"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Using the rdataretriever to quickly analyze Breeding Bird Survey data}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+## Introduction
+
+The Breeding Bird Survey of North America (BBS) is a widely used dataset for understanding geospatial variation and dynamic changes in bird communities.
+The dataset is a continental scale community science project where thousands of birders count birds at locations across North America.
+It has been used in hundreds of research projects including research on biodiversity gradients [@hurlbert2003], ecological forecasting [@harris2019], and bird declines [@wilsey2019].
+
+However working with the Breeding Bird Survey data can be challening because it is composed of roughly 100 different files that need to be cleaned and then combined in multiple ways.
+These initial phases of the data analysis pipeline can require hours of work for even experienced users to understand the detailed layout of the data and either manually assemble it or write code to combine the data.
+The data structure and location also changes regularly meaning that code for this work that is not regularly tested quickly stops working.
+
+This vignette demonstrates how using the `rdataretriever` can eliminate hours of work on data cleaning and restructing allowing researchers to quickly begin addressing interesting scientific questions.
+To do this is demonstrates analyzing the BBS data to evaluate correlates of biodiversity (in the form of species richness).
+
+## Load R Packages
+
+We start by loading the necessary packages.
+In addition to the `rdataretriever` in this demo we'll also use `DBI`, `RSQLite` and `dplyr` to work with the data, `raster` for working with environmental data, and `ggplot2` for visualization.
+
+```{r}
+library(rdataretriever)
+library(DBI)
+library(dplyr)
+library(raster)
+library(ggplot2)
+```
+
+## Install And Connect To The Breeding Bird Survey Data
+
+First we'll update the `rdataretriever` to make sure we have the newest data processing recipes in case something about the structure or location of the dataset has changed.
+Centrally updated data recipes reproducible research with this dataset because something changes every year when the newest data is released meaning that any custom code for processing the data stops working.
+When this happens the `retriever` recipe is updated and after running `get_updates()` data analysis code continues to run as it always has.
+
+```{r, echo = FALSE, results = "hide"}
+rdataretriever::get_updates()
+```
+
+Next install the BBS data into an SQLite database named `bbs.sqlite`.
+
+```{r, echo = FALSE, results = "hide"}
+rdataretriever::install_sqlite('breed-bird-survey', 'bbs.sqlite')
+```
+
+We could also load the data straight into R (using `rdataretriever::fetch('breed-bird-survey')`), store it as flat files (CSV, JSON, or XML), or load it into other database management systems (PostgreSQL, MariaDB, MySQL).
+The data is moderately large (~1GB) so SQLite represents a nice compromise between efficiently conducting the first steps in the data manipulation pipeline while requiring no additional setup or expertise.
+The large number of storage backends makes the `rdataretriever` easy to integrate into existing data processing workflows and to implement designs appropriate to the scale of the data with no additional work. 
+
+Having installed the data into SQLite we can then connect to the database to start analyzing the data.
+
+```{r}
+bbs_db <- dbConnect(RSQLite::SQLite(), 'bbs.sqlite')
+```
+
[…]
