### 2020-10-30 ropensci/rdataretriever#257 @RMHogervorst
Incompatible OSI licence
Hi maintainers/authors,
I noticed the DESCRIPTION files mentions an MIT license, however the LICENSE.md file does not seem to be the MIT license?
