### 2020-10-30 ropensci/rdataretriever#257 @RMHogervorst
I might be mistaken. the text does seem to match, but could you add the header too to make it extremely obvious?
like here: https://mit-license.org/
```
The MIT License (MIT)
```
