### 2020-10-30 70a8b24
Add title to license to make clear it is MIT
