### 2020-10-30 ropensci/rdataretriever#257 @ethanwhite
Thanks @RMHogervorst. You're right that it is the MIT license just without the title. I just submitted a PR to add the title as suggested.
