### 2020-10-30 ropensci/rdataretriever#257 @RMHogervorst
Awesome, and faaaaast!
