### 2020-10-31 ropensci/rdataretriever#259 @RMHogervorst
suggestion remotes::install_github in stead of devtools::install_github
devtools::install_github("ropensci/rdataretriever")
is calling remotes::install_github() underneath.

Devtools has a lot of dependencies that you do not strictly need if you only want to install a package from github.

This is low priority and only a suggestion.
