### 2020-11-01 ropensci/rdataretriever#260 @RMHogervorst
vignette
Hi @henrykironde and others,

First off: this vignette is an amazing overview of the functionality of the package. I think it shows off really nice why you'd need this package. Some details should go in the readme too, but I will open other issues for that.

I noticed the following missing parts:

- Reference file or reference information. Either way the citationkeys are not replaced by their actual citations yet, you see: `[...] gradients [@hurlbert2003]`
- To run the entire vignette you need to have {dbplyr} package installed, automatic tools will not pick that up, so I recommend adding it to the load packages part:

```
library(rdataretriever)
library(DBI)
library(dplyr)
library(dbplyr)
library(raster)
library(ggplot2)
```
