### 2020-11-01 #2800 @RMHogervorst
@henrykironde some notes about the JOSS paper:

The description in `Statement of need` is really beautiful and only after reading that part I finally understood what the problem is that this package is solving. I really really wish you'd include something like that in the Readme (and maybe Description) too.

REQUIRED:
- Add citations so that the references work in the text.
- I am asked to review a `State of the field: Do the authors describe how this software compares to other commonly-used packages?` However I don't see that in the paper at this moment. Could you add something about the state of the field? I don't know what the state of the art is at this moment for the field, but I imagine workflow tools like Drake are used? What is the current state of automation/datapreparation? Is everyone using home brewed scripts? I am genuinely curious.

OPTIONAL / SUGGESTIONS:
- I am not very knowledgeable of the world of ecology or environmental sciences but the concept of frictionless data was new for me. Maybe add one sentence about what that entails?
