### 2020-11-01 #2800 @RMHogervorst
Required Issues opened
- [ ] https://github.com/ropensci/rdataretriever/issues/260  "vignette incomplete"
- [x] https://github.com/ropensci/rdataretriever/issues/257  "clarify OSI licence"

Optional issues opened:
- [ ] https://github.com/ropensci/rdataretriever/issues/259 "suggestion for remotes in stead of devtools::install_github"
