+++ DESCRIPTION
@@ -5,7 +5,7 @@ Description: Provides an R interface to the Data Retriever
     command line interface. The Data Retriever automates the
     tasks of finding, downloading, and cleaning public datasets,
     and then stores them in a local database.
-Version: 3.0.0
+Version: 3.0.1
 Authors@R: c(person("Daniel", "McGlinn", role = "aut",
     email = "danmcglinn@gmail.com",
     comment = c(ORCID = "0000-0003-2359-3526")),
@@ -33,7 +33,7 @@ Authors@R: c(person("Daniel", "McGlinn", role = "aut",
     person("Jeroen", "Ooms", role = "aut",
     email = "jeroenooms@gmail.com",
     comment = c(ORCID = "0000-0002-4035-0289")),
-    person("Francisco Rodríguez-Sánchez", "Ooms", role = "aut",
+    person("Francisco Rodriguez-Sanchez", "Ooms", role = "aut",
     email = "f.rodriguez.sanc@gmail.com",
     comment = c(ORCID = "0000-0002-7981-1599")),
     person("Karthik", "Ram", role = "aut",
@@ -52,7 +52,6 @@ Depends:
     R (>= 3.4.0)
 Imports:
     reticulate (>= 1.16),
-    semver
 Suggests:
     testthat (>= 1.0.0),
     DBI,
