### 2020-11-09 #2800 @henrykironde
Just an update, @RMHogervorst made some great suggestions and I am working to resolve those. Thanks to the issues btw, I guess @jsgalan is inline with the issues raised.
