### 2020-11-12 62e64fb
Provide more explanation of provenance in README
