### 2020-11-12 5acb3b3
Get vignettes building properly with references
Also adds `dbplyr` to BBS vignette to indicate need to install.

Closes #260
