### 2020-11-12 831350a
Provide more explanation of provenance in README (#265)
