### 2020-11-14 ac7a7a7
Use --no-vignettes
