### 2020-11-17 de4fcb6
Add the required Python packages
tables is installed in docker
Remove force arg for hdf5 engine
