+++ R/rdataretriever.R
@@ -517,7 +517,6 @@ install_sqlite <- function(dataset, file = "sqlite.db", table_name = "{db}_{tabl
 #' @param data_dir the dir path to store the db, defaults to working dir
 #' @param debug Setting TRUE helps in debugging in case of errors
 #' @param use_cache Setting FALSE reinstalls scripts even if they are already installed
-#' @param force Setting TRUE doesn't prompt for confirmation while installing committed datasets when changes are discovered in environment
 #' @param hash_value the hash value of committed dataset when installing from provenance directory
 #' @examples
 #' \donttest{
@@ -526,10 +525,10 @@ install_sqlite <- function(dataset, file = "sqlite.db", table_name = "{db}_{tabl
 #' @importFrom reticulate import r_to_py
 #' @export
 install_hdf5 <- function(dataset, file = "hdf5.h5", table_name = "{db}_{table}",
-                           data_dir = getwd(), debug = FALSE, use_cache = TRUE, force = FALSE, hash_value = NULL) {
+                           data_dir = getwd(), debug = FALSE, use_cache = TRUE, hash_value = NULL) {
   tryCatch(withCallingHandlers(
     {
-      retriever$install_hdf5(dataset, file, table_name, data_dir, debug, use_cache, force, hash_value)
+      retriever$install_hdf5(dataset, file, table_name, data_dir, debug, use_cache, hash_value)
     },
     error = function(error_message) {
       message("Full error trace:")
