### 2020-11-17 48c9736
Clear the explanation of commit
#264
