### 2020-11-17 d2a5f29
Emphasize the need and usage of this package to the R-users
fixes #261
