### 2020-11-17 ropensci/rdataretriever#259 @henrykironde
Thanks @RMHogervorst, I will confirm the suggestion with some folks on the team. I encountered and error that was solved by including `devtools`. I will update the issue tomorrow.
