### 2020-11-20 4caf9e0
Edits to explanation of why the rdataretriever is valuable
