### 2020-11-21 #2800 @jsgalan
Hi all,

No problem in the Conda installation

![Screen Shot 2020-11-21 at 10 37 20 AM](https://user-images.githubusercontent.com/971958/99881243-98be1a80-2be6-11eb-9136-1749a2031549.png)

But a problem was found in the R installation.

![Screen Shot 2020-11-21 at 10 43 18 AM](https://user-images.githubusercontent.com/971958/99881246-9b207480-2be6-11eb-88d6-31cb672c44df.png)

Did not found any description on the repository, are there any set requirements for the R installation?

Best
