### 2020-11-21 #2800 @jsgalan
Hi all,

I had problems installing the  spatial datasets

![Screen Shot 2020-11-21 at 1 19 29 PM](https://user-images.githubusercontent.com/971958/99884442-4ab41180-2bfc-11eb-908d-bb8b8df31ccb.png)

Even though I have everything set up in Postgres

![Screen Shot 2020-11-21 at 1 16 41 PM](https://user-images.githubusercontent.com/971958/99884450-56073d00-2bfc-11eb-84d5-6e213bc073cc.png)

Any ideas?

Best
