### 2020-11-22 #2800 @jsgalan
> Thanks @jsgalan for testing this. Make sure you are creating the extensions in the same database that you want to install the data. The default is postgres. In case you, are using a different database, please try to provide the database name in the function call.
> Let me know if that works

Hi all, the extensions are created in all databases. Now I got a problem with finding shp2pgsql

 even thought the bin is listed in the $PATH
