### 2020-11-22 #2800 @jsgalan
Hi @henrykironde

Why are there no references in the article proof?

Am I reading the latest version?

Best
