### 2020-11-22 #2800 @henrykironde
> even thought the bin is listed in the $PATH

Can you try to run `psql` and or `shp2pgsql` from command line

About the citation, I will update that to render correctly
