### 2020-11-22 #2800 @jsgalan
> > even thought the bin is listed in the $PATH
>
> Can you try to run `psql` and or `shp2pgsql` from command line
>
They are accesible from the command line
