+++ paper/paper.bib
@@ -15,3 +15,19 @@
   title = {The EcoData Retriever: Improving Access to Existing Ecological Data},
   journal = {{PLoS} {ONE}}
 }
+
+@Manual{reticulate,
+  author = {JJ Allaire and Kevin Ushey and Yuan Tang and Dirk Eddelbuettel},
+  title = {reticulate: R Interface to Python},
+  year = {2017},
+  url = {https://github.com/rstudio/reticulate},
+}
+
+@Manual{R,
+  title = {R: A Language and Environment for Statistical Computing},
+  author = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address = {Vienna, Austria},
+  year = {2020},
+  url = {https://www.R-project.org/},
+}
