### 2020-11-22 6a2d823
Add reticulate citation (#272)
