+++ paper/paper.md
@@ -88,9 +88,9 @@ Finding, cleaning, standardizing and importing data into efficient data structur
 
 ## Implementation
 
-The main Data Retriever software is written in Python (Morris & White 2013, Senyondo et al 2017). The rdataretriever allows R users to access this data processing workflow through a combination of the reticulate R package (CITATION) and custom features developed for working in R. Because many R users, including the domain researchers most strongly supported by this package, are not familiar with Python and its package management systems, a strong emphasis has been placed on simplifying the installation process for this package so that it can be done entirely from R. Installation requires no direct use of Python or the command line. Detailed documentation has been developed to support users in both installation and use of the software. A Docker based testing systems and associated test suite has also been implemented to ensure that the interoperability of the R package and Python package are maintained, which is challenging due to frequent changes in reticulate and complexities in supporting cross-language functionality across multiple operating systems (Windows, mac os, Linux) and R programming environments (terminal-based R and RStudio).
+The main Data Retriever software is written in Python (Morris & White 2013, Senyondo et al 2017). The rdataretriever allows R users to access this data processing workflow through a combination of the reticulate [@R] package [@reticulate] and custom features developed for working in R. Because many R users, including the domain researchers most strongly supported by this package, are not familiar with Python and its package management systems, a strong emphasis has been placed on simplifying the installation process for this package so that it can be done entirely from R. Installation requires no direct use of Python or the command line. Detailed documentation has been developed to support users in both installation and use of the software. A Docker based testing systems and associated test suite has also been implemented to ensure that the interoperability of the R package and Python package are maintained, which is challenging due to frequent changes in reticulate and complexities in supporting cross-language functionality across multiple operating systems (Windows, mac os, Linux) and R programming environments (terminal-based R and RStudio).
 
-For tabular datasets requiring relatively simple workflows the software uses the JSON based Frictionless Data tabular data metadata package standard (CITATION). For more complex data processing workflows custom Python code is used to process the data into cleaned and standardized formats. Spatial data support is available for PostgreSQL using PostGIS. The information required for handling these datasets is based on a customized version of the Frictionless Data Geo Data schema (CITATION) that also supports raster datasets.
+For tabular datasets requiring relatively simple workflows the software uses the JSON based Frictionless Data tabular data metadata package standard [@frictionlessdata_specs]. For more complex data processing workflows custom Python code is used to process the data into cleaned and standardized formats. Spatial data support is available for PostgreSQL using PostGIS. The information required for handling these datasets is based on a customized version of the Frictionless Data Geo Data schema [@frictionlessdata_specs] that also supports raster datasets.
 
 ## Acknowledgements
 
