### 2020-11-22 #2800 @henrykironde
@jsgalan, I see you are using a Conda environment, make sure that R has access to the same Python paths as well.
Usually, changing to a virtual environment may change the paths. R may be communicating with another python rather that where the programs are installed.

Check the paths in R
```R
library(reticulate)
sys <- import("sys")
sys$path
#Or
reticulate::py_config()
```
Compare to the ones when python is used from the terminal
```Python
import sys
sys.path
```
