### 2020-11-23 #2800 @jsgalan
> @jsgalan, I see you are using a Conda environment, make sure that R has access to the same Python paths as well.
> Usually, changing to a virtual environment may change the paths. R may be communicating with another python rather that where the programs are installed.
>
> Check the paths in R
>
> ```r
> library(reticulate)
> sys <- import("sys")
> sys$path
> #Or
> reticulate::py_config()
> ```
>
> Compare to the ones when python is used from the terminal
>
> ```python
> import sys
> sys.path
> ```

Hi that is pointing to the same place

and
