### 2020-11-23 #2800 @jsgalan
> `reticulate::py_config()` can we try this too
