### 2020-11-23 #2800 @henrykironde
Could you please try this in the terminal
```
import subprocess
subprocess.call('shp2pgsql -d ', shell=True)
```
This must be a path problem but it is hard for me to replicate on my side. Could you try running it on another machine or in a virtual environment.
