### 2020-11-24 #2800 @jsgalan
> Could you please try this in the terminal
>
> ```
> import subprocess
> subprocess.call('shp2pgsql -d ', shell=True)
> ```
>
> This must be a path problem but it is hard for me to replicate on my side. Could you try running it on another machine or in a virtual environment.

Hi,

I think it is doing exactly what I am telling it to do .. 👍

I am finding a way to change it from /sh to /bash where everything is set up. That will probably solve it

Best
