### 2020-11-26 9e00611
Typo fixes
Minor fixes to punctuation and wording
