+++ paper/paper.md
@@ -80,17 +80,17 @@ bibliography: paper.bib
 
 ## Summary
 
-The rdataretriever provides an R interface to the Python-based Data Retriever software. The Data Retriever automates the multiple steps data analysis including downloading, cleaning, and standardizing, and importing tabular datasets and importing them into a variety of relational databases and flat files. It also supports provenance tracking for these steps of the analysis workflow by allowing datasets to be committed at the time of installation and allowing them to be reinstalled with the same data and processing steps in the future. Finally it supports the installation of spatial datasets into relational databases with spatial support. The rdataretriever provides an R interface to all of this functionality, and also supports datasets directly into R for immediate analysis. The system also supports the use of custom data processing routines to support complex datasets that require custom data manipulation steps. The Data Retriever and rdataretriever are focused on scientific data applications including a number of widely used but difficult to work with datasets in ecology and the environmental sciences.
+The rdataretriever provides an R interface to the Python-based Data Retriever software. The Data Retriever automates the multiple steps data analysis including downloading, cleaning, and standardizing, and importing tabular datasets and importing them into a variety of relational databases and flat files. It also supports provenance tracking for these steps of the analysis workflow by allowing datasets to be committed at the time of installation and allowing them to be reinstalled with the same data and processing steps in the future. Finally, it supports the installation of spatial datasets into relational databases with spatial support. The rdataretriever provides an R interface to this functionality and also supports importing of datasets directly into R for immediate analysis. The system also supports the use of custom data processing routines to support complex datasets that require custom data manipulation steps. The Data Retriever and rdataretriever are focused on scientific data applications including a number of widely used, but difficult to work with, datasets in ecology and the environmental sciences.
 
 ## Statement of Need
 
-Finding, cleaning, standardizing and importing data into efficient data structures for modeling and visualization represents a major component of most research workflows. This is a time consuming process for researchers even when working with relatively simple datasets. For more complex datasets these steps can be so complex as to prevent domain experts from engaging with the dataset at all. Systems that operate like package managers for scientific data can overcome these barriers, allowing researchers to move quickly to the final steps in the data analysis workflow (visualization and modeling) and allowing domain experts to leverage the most complex data appropriate to their research questions. The rdataretriever allows R users to automatically conduct these early steps of the analysis workflow for over 200 datasets including a number of the most widely used and difficult to work with datasets in the environmental sciences including the North American Breeding Bird Survey and the Forest Inventory and Analysis datasets. This actively facilitates research on important ecological and environmental questions that would otherwise be limited 
+Finding, cleaning, standardizing and importing data into efficient data structures for modeling and visualization represents a major component of most research workflows. This is a time consuming process for researchers even when working with relatively simple datasets. For more complex datasets these steps can be so complex as to prevent domain experts from engaging with the dataset at all. Systems that operate like package managers for scientific data can overcome these barriers, allowing researchers to move quickly to the final steps in the data analysis workflow (visualization and modeling) and allowing domain experts to leverage the most complex data appropriate to their research questions. The rdataretriever allows R users to automatically conduct these early steps of the analysis workflow for over 200 datasets including a number of the most widely used and difficult to work with datasets in the environmental sciences including the North American Breeding Bird Survey and the Forest Inventory and Analysis datasets. This actively facilitates research on important ecological and environmental questions that would otherwise be limited. 
 
 ## Implementation
 
[…]
