### 2020-11-26 ropensci/rdataretriever#274 @fboehm
Fix citations in paper.md
At least two citations need fixing in paper.md - and perhaps need to be added to paper.bib.

(Morris & White 2013, Senyondo et al 2017)

I suggest formatting these for pandoc, (and adding them to paper.bib, if they aren't there yet) like your other citations in paper.md.
