### 2020-11-26 #2800 @fboehm
@henrykironde - I've opened an issue to address the citations in paper.md. https://github.com/ropensci/rdataretriever/issues/274

Once that is fixed, I'll re-generate the pdf and we can continue.
