### 2020-11-26 #2800 @jsgalan
The written file needs improvement, Please check the highlighted parts below.

[10.21105.joss.02800-3.pdf](https://github.com/openjournals/joss-reviews/files/5604383/10.21105.joss.02800-3.pdf)
