### 2020-11-30 #2800 @jsgalan
Hi all,

When I get to Installing Spatial Datasets I still get error

![Screen Shot 2020-11-29 at 8 36 36 PM](https://user-images.githubusercontent.com/971958/100560269-4f736800-3283-11eb-82ef-d8b4173756be.png)

But it seems that it actually works since it says in the DB schemas
![Screen Shot 2020-11-29 at 8 36 47 PM](https://user-images.githubusercontent.com/971958/100560292-5e5a1a80-3283-11eb-8ad0-9eaa09011609.png)

I confirmed  it was added on Postgres DB

![Screen Shot 2020-11-29 at 8 36 57 PM](https://user-images.githubusercontent.com/971958/100560228-2fdc3f80-3283-11eb-8447-39fd0caa075e.png)

To put my individual issues to rest,  will ask the authors @henrykironde to provide at least a working example of how this connection to the database, and some lines on how to use/show the dataset

I am happy with rest, granted the changes to the article are made.

Best,
