### 2020-11-30 555ed7a
Add missing middle inititials for authors
