+++ paper/paper.bib
@@ -31,3 +31,16 @@
   year = {2020},
   url = {https://www.R-project.org/},
 }
+
+@article{Senyondo2017,
+  doi = {10.21105/joss.00451},
+  url = {https://doi.org/10.21105/joss.00451},
+  year = {2017},
+  publisher = {The Open Journal},
+  volume = {2},
+  number = {19},
+  pages = {451},
+  author = {Henry Senyondo and Benjamin D. Morris and Akash Goel and Andrew Zhang and Akshay Narasimha and Shivam Negi and David J. Harris and Deborah Gertrude Digges and Kapil Kumar and Amritanshu Jain and Kunal Pal and Kevinkumar Amipara and Ethan P. White},
+  title = {Retriever: Data Retrieval Tool},
+  journal = {Journal of Open Source Software}
+}
