### 2020-11-30 #2800 @henrykironde
@jsgalan looks like there may be a problem with the version of python you are using. `psql` is trying to connect using port `5432.0` This should be `5432`. You can try to provide the arguments and see it that will work
`rdataretriever::install_postgres(dataset = "portal", user = "postgres", host = "localhost", port = 5432, database = "postgres")`
