### 2020-12-03 #2800 @jsgalan
> @jsgalan looks like there may be a problem with the version of python you are using. `psql` is trying to connect using port `5432.0` This should be `5432`. You can try to provide the arguments and see it that will work
> `rdataretriever::install_postgres(dataset = "portal", user = "postgres", host = "localhost", port = 5432, database = "postgres")`

Hi all, this worked like a charm!

But, I do not know how to use it, It will be cool to add a small example on how to use this dataset in the **Installing Spatial Datasets** section

Best,
