### 2020-12-05 #2800 @fboehm
@henrykironde - It looks like there are a couple small typos in the paper.md. I think that you should say "A Docker-based testing system" in place of "A Docker based testing systems". That is, change "systems" to "system" and insert a hyphen between "Docker" and "based". Once you make these fixes, we can proceed.
