### 2020-12-06 11a3fd2
Remove typo (#279)
