### 2020-12-06 #2800 @fboehm
@henrykironde - would you make a tagged release on github and archive that release at a service like zenodo? After you've done that, please report the version number and doi in a comment on this thread. Thanks!!
