### 2020-12-22 #2800 @danielskatz
I've proofread the paper and it looks good - it's ready to accept once the zenodo metadata fixes are made.
