### 2021-01-06 #2800 @arfon
@RMHogervorst, @jsgalan - many thanks for your reviews here and to @fboehm for editing this submission. JOSS relies upon the volunteer efforts of folks like yourselves - we simply wouldn't be able to do this without you ✨!
